Moved PDCard to its own class file for organization.
